+++ src/main/java/org/tikv/service/failsafe/MetricsListener.java
@@ -13,17 +13,8 @@
  * limitations under the License.
  */
 
-package org.tikv.common.failsafe;
+package org.tikv.service.failsafe;
 
-import java.io.Closeable;
-
-public interface CircuitBreakerMetrics extends Closeable {
-  /** Record a successful call. */
-  void recordSuccess();
-
-  /** Record a failure call. */
-  void recordFailure();
-
-  /** Add metrics listener. */
-  void addListener(MetricsListener metricsListener);
+public interface MetricsListener {
+  void onNext(HealthCounts healthCounts);
 }
